solution with different components
