+++ src/Components/ErrorMessage.tsx
@@ -1,0 +1,26 @@
+import React, { Dispatch } from 'react';
+
+interface ErrorMessageProps {
+  errorMessage: string;
+  setErrorMessage: Dispatch<React.SetStateAction<string>>;
+}
+
+export const ErrorMessage: React.FC<ErrorMessageProps> = ({
+  errorMessage,
+  setErrorMessage,
+}) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={`notification is-danger is-light has-text-weight-normal ${!errorMessage.trim() ? 'hidden' : ''}`}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={() => setErrorMessage('')}
+      />
+      {errorMessage}
+    </div>
+  );
+};
