+++ src/Components/Footer.tsx
@@ -1,0 +1,121 @@
+import React, { Dispatch, RefObject } from 'react';
+import { Filter, Todo } from '../types/Todo';
+
+interface FooterProps {
+  setLoadingTodoId: Dispatch<React.SetStateAction<number[] | null>>;
+  deleteTodo: (todoId: number) => Promise<void>;
+  setErrorMessage: Dispatch<React.SetStateAction<string>>;
+  todos: Todo[];
+  setTodos: Dispatch<React.SetStateAction<Todo[]>>;
+  inputRef: RefObject<HTMLInputElement>;
+  filter: Filter;
+  setFilter: Dispatch<React.SetStateAction<Filter>>;
+}
+
+export const Footer: React.FC<FooterProps> = ({
+  setLoadingTodoId,
+  deleteTodo,
+  setErrorMessage,
+  todos,
+  setTodos,
+  inputRef,
+  filter,
+  setFilter,
+}) => {
+  function handleDeleteAllCompletedTodos(completedTodos: Todo[]) {
+    const completedIds = completedTodos.map(todo => todo.id);
+
+    setLoadingTodoId(completedIds);
+
+    const deletePromises = completedTodos.map(todo =>
+      deleteTodo(todo.id)
+        .then(() => todo.id)
+        .catch(() => {
+          setErrorMessage('Unable to delete a todo');
+
+          return Promise.reject(todo.id);
+        }),
+    );
+
+    Promise.allSettled(deletePromises)
+      .then(results => {
+        const successfulIds = results
+          .filter(result => result.status === 'fulfilled')
+          .map(result => (result as PromiseFulfilledResult<number>).value);
+
+        const failedIds = results
+          .filter(result => result.status === 'rejected')
+          .map((_, index) => completedIds[index]);
+
+        setTodos(prevTodos =>
+          prevTodos.filter(todo => !successfulIds.includes(todo.id)),
+        );
+
+        if (failedIds.length > 0) {
+          setErrorMessage('Unable to delete some todos');
+        }
+      })
+      .catch(() => {
+        setErrorMessage('Unable to delete a todo');
+      })
+      .finally(() => {
+        setLoadingTodoId(null);
+        setTimeout(() => {
+          if (inputRef.current) {
+            inputRef.current.focus();
+          }
+        }, 0);
+      });
+  }
+
+  return (
+    <footer className="todoapp__footer" data-cy="Footer">
+      <span className="todo-count" data-cy="TodosCounter">
+        {`${todos.filter(todo => !todo.completed && todo.id !== -1).length} items left`}
+      </span>
+
+      <nav className="filter" data-cy="Filter">
+        <a
+          href="#/"
+          className={`filter__link ${filter === Filter.All ? 'selected' : ''}`}
+          data-cy="FilterLinkAll"
+          onClick={() => setFilter(Filter.All)}
+        >
+          All
+        </a>
+
+        <a
+          href="#/active"
+          className={`filter__link ${filter === Filter.Active ? 'selected' : ''}`}
+          data-cy="FilterLinkActive"
+          onClick={() => setFilter(Filter.Active)}
+        >
+          Active
+        </a>
+
+        <a
+          href="#/completed"
+          className={`filter__link ${filter === Filter.Completed ? 'selected' : ''}`}
+          data-cy="FilterLinkCompleted"
+          onClick={() => setFilter(Filter.Completed)}
+        >
+          Completed
+        </a>
+      </nav>
+
+      <button
+        type="button"
+        className="todoapp__clear-completed"
+        data-cy="ClearCompletedButton"
+        disabled={!todos.some(todo => todo.completed)}
+        onClick={() => {
+          const completedTodos = todos.filter(todo => todo.completed);
+
+          handleDeleteAllCompletedTodos(completedTodos);
+        }}
+      >
+        Clear completed
+      </button>
+    </footer>
+  );
+};
